feat: add user analytics card and table views with modal for detailed insights
